blanket trait to gen record from repo
