+++ src/lib.rs
@@ -7,6 +7,7 @@ pub use crate::repo::DirMeta;
 pub use crate::repo::Entry;
 pub use crate::repo::Hasher;
 pub use crate::repo::Repository;
+pub use crate::repo::RepositoryRecord;
 
 pub mod repo_impl;
 
@@ -14,7 +15,7 @@ mod helper;
 pub use crate::helper::json_extract;
 
 mod resolver;
-pub use crate::resolver::{resolve, RepositoryRecord};
+pub use crate::resolver::resolve;
 
 pub mod crawler;
 pub use crawler::crawl;
